+++ lib/onlyoffice_digitalocean_wrapper/digitalocean_wrapper.rb
@@ -6,6 +6,7 @@ require_relative 'digitalocean_wrapper/exceptions_retryer'
 require_relative 'digitalocean_wrapper/getters'
 require_relative 'digitalocean_wrapper/logger_wrapper'
 require_relative 'digitalocean_wrapper/power_actions'
+require_relative 'digitalocean_wrapper/ssh_helper'
 require_relative 'digitalocean_wrapper/token_methods'
 
 # Namespace for this gem
@@ -16,6 +17,7 @@ module OnlyofficeDigitaloceanWrapper
     include LoggerWrapper
     include ExceptionsRetryer
     include PowerActions
+    include SshHelper
     include TokenMethods
 
     # @return [Array<String>] list of allowed droplet sizes